src/requirements.txt
howlongtobeatpy
requests
python-dotenv
halo